[LibOS] Split off x86_64 specifics from shim_types.h
Split off x86_64 specifics from shim_types.h and move them
into include/arch/x86_64/shim_types-arch.h.
